requirements.txt
langchain
google-generativeai
langchain-google-genai
streamlit
pymupdf
python-dotenv
tiktoken
faiss-cpu